requirements.txt
speechrecognition
pyttsx3
opencv-python
pillow
pyaudio
